# add benchmarks for multivalued fastfield merge