@PSeitz 2022-09-22
It's `SortedDocIdMultiValueColumn::get_val` via Line::train